style: 온보딩 화면 자연스러운 전환
